Move Markdown note type into its own type_widgets/markdown folder
Relocate the Markdown editor out of the shared `code` folder into a dedicated
`markdown` folder, and drop the now-redundant `markdown_` filename prefix
(markdown_completions -> completions, markdown_editor_utils -> editor_utils).
The shared code-snippet helpers stay in `code/snippets`; completions.ts now
imports them via ../code/snippets. Updates the note-type registry and the
llm_chat MarkdownCommons.css import accordingly.
